debug

+++ Alignment/evaluation/eval_rewardbench.py
@@ -147,8 +147,8 @@ def main():
             score_rejected_batch = [result["score"] for result in rewards_rejected]
         # for classes that directly output scores (custom code)
         else:
-            score_chosen_batch = rewards_chosen.cpu().numpy().tolist()
-            score_rejected_batch = rewards_rejected.cpu().numpy().tolist()
+            score_chosen_batch = rewards_chosen.to(torch.float32).cpu().numpy().tolist()
+            score_rejected_batch = rewards_rejected.to(torch.float32).cpu().numpy().tolist()
 
         # log results
         [
